langchain_demo/requirements.txt
# core dependencies
snowflake-snowpark-python

# langchain dependencies
langchain==0.1.16
sentence_transformers
torch
